changed it so option to update already imported events is per feed instead of global

+++ includes/ical/class-calendar-plus-ical-sync.php
@@ -35,7 +35,7 @@ class Calendar_Plus_iCal_Sync {
 	 * Update events or keep them original
 	 * @var bool
 	 */
-	protected $update_events;
+	protected $keep_updated;
 
 	/**
 	 * Parent feed url
@@ -56,14 +56,14 @@ class Calendar_Plus_iCal_Sync {
 			'author'        => 0,
 			'category'      => 0,
 			'status'        => 'publish',
-			'update_events' => false,
+			'keep_updated' => 0,
 			'source'        => '',
 		] );
 
-		$this->event_author = $args['author'] ? $args['author'] : get_current_user_id();
+		$this->event_author   = $args['author'] ? $args['author'] : get_current_user_id();
 		$this->event_category = $args['category'];
 		$this->default_status = $args['status'];
-		$this->update_events  = $args['update_events'];
+		$this->keep_updated   = $args['keep_updated'];
 		$this->feed_url       = $args['source'];
 	}
 
@@ -120,22 +120,29 @@ class Calendar_Plus_iCal_Sync {
 			'post_content' => $event_data['post_content'],
 		);
 
+		update_post_meta( $event->ID, '_event_feed_url', $this->feed_url );
+
 		if ( $event ) {
+			if ( ! $this->keep_updated ) {
+				return false;
+			}
+
 			if ( 'trash' === $event_data['post_status'] ) {
 				// The feed event has been deleted, delete locally
 				wp_delete_post( $event->ID, true );
 				return $event->ID;
 			}
 
-			update_post_meta( $event->ID, '_event_feed_url', $this->feed_url );
-
-			if ( ! $this->update_events ) {
-				return false;
-			}
-
 			$old_event_hash = $event->get_meta( 'ical_hash' );
-			if ( $old_event_hash === $event_data_hash ) {
-				return false;
+			if( $old_event_hash ) {
+				if ( $old_event_hash === $event_data_hash ) {
+					return false;
+				}
+			}
+			else {
+				if ( ! $event_data['last_updated'] || $event_data['last_updated'] === $event->get_meta( 'ical_last_updated' ) ) {
+					return false;
+				}
 			}
 
 			$post_args['ID'] = $event->ID;
@@ -146,15 +153,12 @@ class Calendar_Plus_iCal_Sync {
 
 			wp_update_post( $post_args );
 			$post_id = $event->ID;
-
 		} else {
 			if ( 'trash' === $event_data['post_status'] ) {
 				return false;
 			}
 
 			$post_id = wp_insert_post( $post_args );
-
-			update_post_meta( $post_id, '_event_feed_url', $this->feed_url );
 		}
 
 		update_post_meta( $post_id, '_event_uid', $event_data['uid'] );
